% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneInfiltration.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{CIBERSORT}
\usage{
CIBERSORT(
  sig_matrix,
  mixture_file,
  perm = 100,
  QN = TRUE,
  sigMat = "LM22",
  sample = "multiple"
)
}
\arguments{
\item{sig_matrix}{Import signature matrix data,'LM22' or 'LM14'.}

\item{mixture_file}{User upload data.}

\item{perm}{Number of permutations.}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE).}

\item{sigMat}{signature matrix name for choose,'LM22' or 'LM14'.}

\item{sample}{Choose 'multiple' or 'single' samples.}

\item{saveDir}{The storage path of results.}
}
\description{
The function is referenced by CIBERSORT_server to produce other files.

Download this file from https://cibersort.stanford.edu/
CIBERSORT R script v1.03 (last updated 07-10-2015)
Primary Author: Aaron M. Newman, Stanford University (amnewman@stanford.edu)

There is a little change compared to the source code from CIBERSORT, and we already obtain permission.
License: http://cibersort.stanford.edu/CIBERSORT_License.txt
}
\details{
There are some change in parameters for our use , we need to choose diffrent signature matrix, deal with single
and multiple samples.
}
