% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneInfiltration.R
\name{CIBERSORT.server}
\alias{CIBERSORT.server}
\alias{CIBERSORT}
\title{Immune cell infiltration proportion calculate module.}
\usage{
CIBERSORT.server(
  data,
  perm = 100,
  CHIPorRNASEQ = "RNA-seq",
  sample = "multiple",
  dataType = "TPM"
)
}
\arguments{
\item{data}{data}

\item{perm：}{The NO. permutation. Default by 100.}

\item{CHIPorRNASEQ:}{The source of data,'RNAseq' or 'Microarray' to choose. Default by "RNA-seq".}

\item{sample：}{The number of upload sample, "multiple" or "single" sample for user to choose. Default by "multiple".}

\item{dataType:}{The data type of upload data, "count" or "TPM" for user to choose. Default by "TPM".}
}
\description{
A reference function to produce the calculate cell infiltration proportions results
and interactive picture data of the TIP website.
}
\author{
DengChunyu, Haerbin Medical University, dengcyelena@gmail.com
Requirements:
      R v3.0 or later. (dependencies below might not work properly with earlier versions)
      install.packages('e1071')
      install.packages('parallel')
      install.packages('preprocessCore')
      If preprocessCore is not available in the repositories you have selected, run the following:
          source("http://bioconductor.org/biocLite.R")
          biocLite("preprocessCore")
Windows users using the R GUI may need to Run as Administrator to install or update packages.
}
