% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneInfiltration.R
\name{CIBERSORT_server}
\alias{CIBERSORT_server}
\title{The web server file about immune infiltration}
\usage{
CIBERSORT_server(
  mixture_file,
  sig_matrix,
  perm = 100,
  QN = TRUE,
  CHIPorRNASEQ = "RNA-seq",
  sample = "multiple"
)
}
\arguments{
\item{mixture_file}{Input data of user upload.}

\item{sig_matrix}{Input data of signature matrix.}

\item{perm}{The NO. permutations,100 by defaut.}

\item{QN}{TRUE by defaut ,Perform quantile normalization or not.}

\item{CHIPorRNASEQ}{The mixture_file are "Microarray" data or "RNA-seq" data, for user to choose.}

\item{sample}{User choose 'single' and 'multiple' samples.}
}
\description{
This function is to produce all the results file used to produce interactive graphics by web server.
We reference the result of function 'CIBERSORT' to produce graphics files of web server need. Including
radar plot,stackplot and pie plot.
}
\details{
The interactive graphics files are used to web to produce beautiful and powerful interactive graphics. The format
of these file is specific for graphics.
}
\author{
DengChunyu
}
