% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2Tpm.R
\name{count2Tpm}
\alias{count2Tpm}
\title{Transform the data type of RNA-seq from count to TPM}
\usage{
count2Tpm(countName, filePath, sample = "multiple")
}
\arguments{
\item{countName}{The name of counts file.}

\item{filePath}{The storage path of 'genename_length2.0.Rdata'.}

\item{sample}{Choose "multiple" or 'single' for sample number.}
}
\description{
Convert a numeric matrix of features (rows) and conditions (columns) with
raw feature counts to transcripts per million(TPM).
}
\details{
We need 'genename_length2.0.Rdata' to caculate TPM data.
}
