% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOutliers.R
\name{filterOutliers}
\alias{filterOutliers}
\title{@ filterOutliers}
\usage{
filterOutliers(raw.matrix)
}
\arguments{
\item{raw.matrix}{A matrix consist of numeric values.}
}
\value{
raw matrix without outliers
}
\description{
Filter outliers of a matrix according to the criteria of the boxplot.
}
