% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.core.R
\name{indicatorInsideGeneSetMatrix}
\alias{indicatorInsideGeneSetMatrix}
\title{@ indicatorInsideGeneSetMatrix}
\usage{
indicatorInsideGeneSetMatrix(superposition.rank, gSet.position, m, n)
}
\arguments{
\item{superposition.rank}{A matrix indicating rank of all values in the raw matrix.}

\item{gSet.position}{A matrix consist of '0' and '1' show the position of signature genes in expression profile, 
with same size of expression profile, '1' stands for signature genes expression, '0' for other genes.}

\item{m}{Numeric value indicating number of row of matrix 'superposition.rank'.}

\item{n}{Numeric value indicating number of column of matrix 'superposition.rank'.}
}
\description{
With the rank of all values in the matrix(superposition.rank), 
get the matrix consist with 0 and 1, indicated the position of geneset, required for calculating ES score of the geneset.
}
