% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.core.R
\name{oneSetssGSEA.ES}
\alias{oneSetssGSEA.ES}
\title{@ oneSetssGSEA.ES}
\usage{
oneSetssGSEA.ES(
  exp.rank,
  superposition.rank.es,
  gSet.position.matrix,
  alpha = 0.25,
  m.num,
  n.num
)
}
\arguments{
\item{exp.rank}{A matrix indicating rank of every element in column.}

\item{superposition.rank.es}{A matrix indicating rank of all values in the matrix.}

\item{gSet.position.matrix}{A matrix consist of '0' and '1' show the position of signature genes in expression profile, 
with same size of expression profile, '1' stands for signature genes expression, '0' for other genes.}

\item{m.num}{Numeric value indicating number of row of matrix 'exp.rank'.}

\item{n.num}{Numeric value indicating number of column of matrix 'exp.rank'.}

\item{alpha=0.25}{Quantity alpha usually set to 0.25.}
}
\value{
vector of numeric values indicating ES score of one geneset for every sample
}
\description{
Calculating ES score of one geneset.
}
\author{
Liwen Xu
}
