% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessMultipleSample.R
\name{ProcessMultipleSample}
\alias{ProcessMultipleSample}
\title{@ ProcessMultipleSample}
\usage{
ProcessMultipleSample(
  expression.from.users,
  signatureList = signature.GeneSymbol.list,
  perm.times = 100,
  signature.annotation = annotation,
  type.of.data = c("Microarray", "RNA-seq"),
  format.of.file = c("TPM", "Count"),
  gene.length.path
)
}
\arguments{
\item{expression.from.users}{A data.frame contains expression data from users.}

\item{signatureList}{List of signature genes, default by 'signature.GeneSymbol.list'.}

\item{perm.times}{Numeric value indicating times of permutation, default by 100.}

\item{signature.annotation}{A data.frame contains step and function direction information about signature genes,
default by 'annotation'.}

\item{type.of.data}{Character indicating source of expression data, 'Microarray'or 'RNA-seq'.}

\item{format.of.file}{Character indicating format of RNA-seq expression data, 'TPM' or 'Count'.}

\item{gene.length.path}{Character represting the storage path of the gene length file.}
}
\description{
Several steps to deal with users uploaded multi-sample expression profile:
calculating and normalizing ES score of samples, combine the normalized ES score from
positive set and negative set of genes belong to the same step by calculating the difference between them,
then change the format of results for visualization on web.
}
