% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.core.R
\name{rankMatrixByCol}
\alias{rankMatrixByCol}
\title{@ rankMatrixByCol}
\usage{
rankMatrixByCol(superposition.rank, m, n)
}
\arguments{
\item{superposition.rank}{A matrix indicating rank of all values in the raw matrix.}

\item{m}{Numeric value indicating number of row of matrix 'superposition.rank'.}

\item{n}{Numeric value indicating number of column of matrix 'superposition.rank'.}
}
\value{
rank of elements in every column
}
\description{
With the rank of all values in the matrix(superposition.rank), restore the rank of elements in every column.
}
