% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.core.R
\name{sortMatrixByCol}
\alias{sortMatrixByCol}
\title{@ sortMatrixByCol}
\usage{
sortMatrixByCol(raw.matrix, set.decreasing = TRUE)
}
\arguments{
\item{raw.matrix}{A matrix consist of numeric value.}

\item{set.decreasing}{Logical, if true sort is decreasing, default by TRUE.}
}
\description{
Sort every column of a matrix.
}
