% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEAPermutation.R
\name{ssGSEA.MultipleSamples.Permutation}
\alias{ssGSEA.MultipleSamples.Permutation}
\title{@ ssGSEA.MultipleSamples.Permutation}
\usage{
ssGSEA.MultipleSamples.Permutation(exp.profile)
}
\arguments{
\item{exp.profile}{A data.frame indicating the real expression profile.}
}
\value{
data.frame of random expression profile
}
\description{
Perturb the rownames of an expression profile in order to obtain a random expression profile.
}
