% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea.core.R
\name{superpositionRank}
\alias{superpositionRank}
\title{@ superpositionRank}
\usage{
superpositionRank(raw.matrix, m, n)
}
\arguments{
\item{raw.matrix}{A matrix consist of numeric values.}

\item{m}{Numeric value indicating number of row of raw.matrix.}

\item{n}{Numeric value indicating number of column of raw.matrix.}
}
\value{
rank of all values in the matrix
}
\description{
Get rank of a matrix based on superposition by gradient. 
Required for getting rank of elements in every column of a matrix.
}
